feature: fine grained modules for core
