+++ packages/core/tsup.config.ts
@@ -4,18 +4,34 @@ export default defineConfig({
   ...packageOptions,
   config: "tsconfig.build.json",
   entry: [
+    "./src/eventFilterMatches.ts",
     "./src/index.ts",
+    "./src/replacements.ts",
+    "./src/requestFilterMatches.ts",
+    "./src/retry.ts",
+    "./src/schemas/index.ts",
+    "./src/types.ts",
+    "./src/utils.ts",
+    "./src/versions.ts",
+    "./src/v3/dev/index.ts",
+    "./src/v3/errors.ts",
     "./src/v3/index.ts",
+    "./src/v3/logger-api.ts",
     "./src/v3/otel/index.ts",
+    "./src/v3/prod/index.ts",
+    "./src/v3/schemas/index.ts",
+    "./src/v3/semanticInternalAttributes.ts",
+    "./src/v3/utils/durations.ts",
+    "./src/v3/utils/flattenAttributes.ts",
+    "./src/v3/utils/ioSerialization.ts",
+    "./src/v3/utils/omit.ts",
+    "./src/v3/utils/retries.ts",
+    "./src/v3/utils/structuredLogger.ts",
+    "./src/v3/workers/index.ts",
+    "./src/v3/zodfetch.ts",
+    "./src/v3/zodIpc.ts",
     "./src/v3/zodMessageHandler.ts",
     "./src/v3/zodNamespace.ts",
     "./src/v3/zodSocket.ts",
-    "./src/v3/zodIpc.ts",
-    "./src/v3/utils/structuredLogger.ts",
-    "./src/v3/utils/durations.ts",
-    "./src/v3/dev/index.ts",
-    "./src/v3/prod/index.ts",
-    "./src/v3/workers/index.ts",
-    "./src/v3/zodfetch.ts",
   ],
 });
